removed the traffic pattern obfuscation

+++ ll/onionstew/attachsc.go
@@ -74,17 +74,21 @@ func (ctx *sc_ctx) AttachSC(wire io.ReadWriteCloser, serverside bool) {
 				ctx.destroy()
 				// Will die on next iteration
 			}
-			xaxa := make([]byte, 1)
-			cryptorand.Read(xaxa)
-			for len(newthing.payload) < 1024 && xaxa[0] < 128 {
+
+			//TODO: Make this not tie up the connection
+			if false {
+				xaxa := make([]byte, 1)
 				cryptorand.Read(xaxa)
-				xaxa := sc_message{0xFFFFFFFFFFFFFFFE, make([]byte, xaxa[0])}
-				err := write_sc_message(xaxa, wire)
-				if err != nil {
-					kilog.Warning("AttachSC encountered unexpected error %s while WRITING KA, DESTROYING STEW",
-						err.Error())
-					ctx.destroy()
-					// Will die on next iteration
+				for len(newthing.payload) < 1024 && xaxa[0] < 128 {
+					cryptorand.Read(xaxa)
+					xaxa := sc_message{0xFFFFFFFFFFFFFFFE, make([]byte, xaxa[0])}
+					err := write_sc_message(xaxa, wire)
+					if err != nil {
+						kilog.Warning("AttachSC encountered unexpected error %s while WRITING KA, DESTROYING STEW",
+							err.Error())
+						ctx.destroy()
+						// Will die on next iteration
+					}
 				}
 			}
 		case <-local_stop:
